Use base class for searchIndex tests
Many tests are the same for File and Memory indexes
