+++ _test/tests/inc/Search/Collection/FullTextCollectionTest.php
